[FIX] Searching hidden user info
Has now been fixed for the following cards:
- Manager
- Projects
- Education
- Classification
- Career Suummary
- Acting role

NOTE:
Searching for the following information does not work. Future bug to be fixed:
- Competencies
- Talent management
- Developmental goals
- Career interests
I suspect this is due to the fact that we're using the visiba cards profile info in the API/profile/get file when it is most likely undefined in at that point.